code health: move CollectAppList function from pack to util package
Follow up #18
