+++ cli/util/util_test.go
@@ -15,6 +15,7 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/tarantool/tt/cli/cmdcontext"
 	"github.com/tarantool/tt/cli/config"
+	"github.com/tarantool/tt/cli/pack/test_helpers"
 )
 
 type inputValue struct {
@@ -483,5 +484,51 @@ func TestInstantiateFileFromTemplate(t *testing.T) {
 			require.NoError(t, err)
 			assert.Equal(t, tt.expectedContent, string(content))
 		})
+	}
+}
+
+func TestCollectAppList(t *testing.T) {
+	testDir := t.TempDir()
+	var defaultPaths = []string{
+		"var",
+		"log",
+		"run",
+		"lib",
+		"env",
+		filepath.Join("env", "bin"),
+		filepath.Join("env", "modules"),
+	}
+
+	apps := map[string]bool{
+		"app1.lua": true,
+		"app2":     true,
+	}
+
+	dirsToCreate := []string{
+		"app2",
+		".rocks",
+	}
+	dirsToCreate = append(dirsToCreate, defaultPaths...)
+
+	filesToCreate := []string{
+		"app1.lua",
+		"somefile",
+		"app2/init.lua",
+	}
+
+	err := test_helpers.CreateDirs(testDir, dirsToCreate)
+	require.NoErrorf(t, err, "failed to initialize a directory structure: %v", err)
+
+	err = test_helpers.CreateFiles(testDir, filesToCreate)
+	require.NoErrorf(t, err, "failed to initialize a directory structure: %v", err)
+
+	collected, err := CollectAppList(testDir)
+	assert.Nilf(t, err, "failed to collect an app list: %v", err)
+
+	require.Equalf(t, len(apps), len(collected), "wrong count applications collected,"+
+		" expected: %d, got %d", len(apps), len(collected))
+
+	for _, item := range collected {
+		require.Truef(t, apps[item], "wrong item got collected in app list: %s", item)
 	}
 }
